Add maths vars module.
Register a variable at a given address with a linear maths function so v=a+b*f(c + d*i) where i is an iteration value incremented each tick.
This can be used to drive variables independently of other code and controlled by the sequence script.
cf. Rocket but with linear functions.
In theory could chain these together as the order of evaluation is guaranteed to be the order the variables are added to the list.